[FEATURE] Add "update" mode
Adds support for a new "mode" option value: "update".

If mode=update is set, the script will check if the entry in YAML
is defined with a "uid" property. If a "uid" is defined, the script
checks if the corresponding record can be found - and if found,
the script uses SQL update to update the record. If not found,
the record is inserted (like it happens with mode=append).

This allows entries to be updated without a truncate occurring
as it does with mode=replace. In turn it allows multiple sets of
data for the same table to be maintained (with forced "uid").