[BUGFIX] Ignore empty data sets
Ignores data sets which don't have any "entries", which
in turn allows a data set to define mode=replace to
cause a truncate of the table without adding any new
records to it.

Useful in cases where several partial data sets are added
with mode=update, but where an initial truncate is wanted
in order to clear the table before processing the partial
data sets.